Added all equations for activity coefficients
Currently, equations for activity coefficients are only available
as gamma1(T, x1, parameters) and no inverse functions is implemented.
Moreover, volumes of pure components v1 and v2, which are required
for some equations like Wilson, are assumed to be constant and to
be defined in the JSON file. Therefore, I will add a more generic
function that allows the user to either use the constant values
given in the JSON file or to give the volumes as inputs.

This is for issue #1.